riscv: Fix RISCV_ALTERNATIVE_EARLY
mainline inclusion
from mainline-6.11-rc7
commit 1ff95eb2bebda50c4c5406caaf201e0fcb24cc8f
category: feature
bugzilla: https://github.com/RVCK-Project/rvck/issues/239

--------------------------------

RISCV_ALTERNATIVE_EARLY will issue sbi_ecall() very early in the boot
process, before the first memory mapping is setup so we can't have any
instrumentation happening here.

In addition, when the kernel is relocatable, we must also not issue any
relocation this early since they would have been patched virtually only.

So, instead of disabling instrumentation for the whole kernel/sbi.c file
and compiling it with -fno-pie, simply move __sbi_ecall() and
__sbi_base_ecall() into their own file where this is fixed.

Reported-by: Conor Dooley <conor.dooley@microchip.com>
Closes: https://lore.kernel.org/linux-riscv/20240813-pony-truck-3e7a83e9759e@spud/
Reported-by: syzbot+cfbcb82adf6d7279fd35@syzkaller.appspotmail.com
Closes: https://lore.kernel.org/linux-riscv/00000000000065062c061fcec37b@google.com/
Fixes: 1745cfafebdf ("riscv: don't use global static vars to store alternative data")
Link: https://lore.kernel.org/r/20240829165048.49756-1-alexghiti@rivosinc.com
